requirements.txt
django==1.11.17
gunicorn==19.8.1
xadmin==0.6.1
mysqlclient==2.0.3
django-ckeditor==5.6.1
djangorestframework==3.10.0
django-redis==4.7.0
django-autocomplete-light==3.2.10
Markdown==3.3.7
Pillow==5.3.0
coreapi==2.3.3
hiredis==2.0.0
django-simple-captcha==0.5.10
raven==6.9.0
captcha==0.2.4
supervisor==4.2.4
